# fix: Improve `cast` performance on macos - Only load root TLS certificates with HTTPs rpc-url