@RedaOps 2023-11-18
This change will be deprecated once the alloy providers are _fully_ implemented since the `RuntimeClient` will be removed. Maybe this is something that can be ported to the new providers, or do they use a different client?