@mattsse 2023-11-18
> This change will be deprecated once the alloy providers are fully implemented since the RuntimeClient will be removed

they' still be running reqwest so we def want that fix