@drortirosh 2023-11-18
was this supposed to improve the performance of cast? because it didn't.
on latest foundry () I get:
```sh
 $ time cast ci
31337

real	0m0.199s    <<< curl does this at <17ms
user	0m0.097s
sys	0m0.034s
```