@RedaOps 2023-11-18
> was this supposed to improve the performance of cast? because it didn't.
> 
> on latest foundry (`cast 0.2.0 (c948388 2023-11-18T00:19:39.244711000Z`) on MacOs I get:
> 
> ```sh
> 
>  $ time cast ci
> 
> 31337
> 
> 
> 
> real	0m0.199s    <<< curl does this at <17ms
> 
> user	0m0.097s
> 
> sys	0m0.034s
> 
> ```
> 
> 

You are running the version right before this was merged 😅

Commit c948388a0768eb0017dd579b1ad1d8b4f33627c4 (which you are running) is the commit right before the fix.

Can you update it to the new master and try again?