@DaniPopes 2023-11-18
@drortirosh Versions are released on a **nightly** basis, so either build off of master branch with `foundryup -b master` or wait until tomorrow.

@RedaOps Awesome work btw :D